+++ src/Engine/Font/FontFamily.cpp
@@ -2,7 +2,7 @@
 
 using namespace Engine;
 
-FontFamily::FontFamily(VulkanRenderer& vulkan, const Path& dir, const std::string& name, const float size) :
+FontFamily::FontFamily(VulkanRenderer& vulkan, const Path& dir, const std::string& name, const int size) :
     faces{
         FontFace{vulkan, dir / Path{name + "-regular.ttf"}, size},
         FontFace{vulkan, dir / Path{name + "-bold.ttf"}, size},
